// apps/web/src/lib/api/generated/model/renewalColumnStage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type RenewalColumnStage = typeof RenewalColumnStage[keyof typeof RenewalColumnStage];


export const RenewalColumnStage = {
  due: 'due',
  contacted: 'contacted',
  quoted: 'quoted',
  renewed: 'renewed',
  lost: 'lost',
} as const;